Relocate component to shared dir
