+++ contentcuration/contentcuration/frontend/shared/views/RecommendedResourceCard.vue
